+++ Classes/Model/Checkbox.php
@@ -5,12 +5,11 @@ declare(strict_types=1);
 namespace OliverKlee\Seminars\Model;
 
 use OliverKlee\Oelib\Model\AbstractModel;
-use OliverKlee\Seminars\Model\Interfaces\Titled;
 
 /**
  * This class represents a checkbox.
  */
-class Checkbox extends AbstractModel implements Titled
+class Checkbox extends AbstractModel
 {
     /**
      * @return string our title, will not be empty