+++ Classes/Model/EventType.php
@@ -5,12 +5,11 @@ declare(strict_types=1);
 namespace OliverKlee\Seminars\Model;
 
 use OliverKlee\Oelib\Model\AbstractModel;
-use OliverKlee\Seminars\Model\Interfaces\Titled;
 
 /**
  * This class represents an event type.
  */
-class EventType extends AbstractModel implements Titled
+class EventType extends AbstractModel
 {
     /**
      * @return string our title, will not be empty
@@ -18,18 +17,6 @@ class EventType extends AbstractModel implements Titled
     public function getTitle(): string
     {
         return $this->getAsString('title');
-    }
-
-    /**
-     * @param string $title our title to set, must not be empty
-     */
-    public function setTitle(string $title): void
-    {
-        if ($title == '') {
-            throw new \InvalidArgumentException('The parameter $title must not be empty.', 1333296812);
-        }
-
-        $this->setAsString('title', $title);
     }
 
     /**