+++ Classes/Model/TargetGroup.php
@@ -5,12 +5,11 @@ declare(strict_types=1);
 namespace OliverKlee\Seminars\Model;
 
 use OliverKlee\Oelib\Model\AbstractModel;
-use OliverKlee\Seminars\Model\Interfaces\Titled;
 
 /**
  * This class represents a target group.
  */
-class TargetGroup extends AbstractModel implements Titled
+class TargetGroup extends AbstractModel
 {
     /**
      * @return string our title, will not be empty