+++ Tests/LegacyUnit/Model/RegistrationTest.php
@@ -51,34 +51,6 @@ final class RegistrationTest extends TestCase
     /**
      * @test
      */
-    public function setTitleWithEmptyTitleThrowsException(): void
-    {
-        $this->expectException(
-            \InvalidArgumentException::class
-        );
-        $this->expectExceptionMessage(
-            'The parameter $title must not be empty.'
-        );
-
-        $this->subject->setTitle('');
-    }
-
-    /**
-     * @test
-     */
-    public function setTitleSetsTitle(): void
-    {
-        $this->subject->setTitle('registration for event');
-
-        self::assertEquals(
-            'registration for event',
-            $this->subject->getTitle()
-        );
-    }
-
-    /**
-     * @test
-     */
     public function getTitleWithNonEmptyTitleReturnsTitle(): void
     {
         $this->subject->setData(['title' => 'registration for event']);